+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## CHANGELOG
 
 ### UNRELEASED CHANGES
+- Documentation: update README.
+- Add: Support using a library registry in the CONFIG_FILE so that the community deployment can use the SimplyE library registry to display all libraries. 
 
 - Add: Add catalog button on openE home page
 - Fix: Protect against invalid redirect uris caused by mismatch in hydration.